remove linting file and add linting function to file_store
